+++ Documentation/Allen_Doc_Gen.lua
@@ -1,7 +1,7 @@
 --------------------------------------------------------------------------
 -- Set of utility functions for strings
 -- @author Roland Yonaba
--- @release $Id: Allen.lua,v 0.1 07/24/2012 Roland_Yonaba$
+-- @release $Id: Allen.lua,v1.0 08/02/2012 Roland_Yonaba$
 --------------------------------------------------------------------------
 
 --Copyright (c) 2012 Roland Yonaba
@@ -27,77 +27,6 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 --]]
 
 ------------------------------------------------------------------
--- Internalization
-------------------------------------------------------------------
-
-local _ = {}
-local string = string
-local pairs,ipairs = pairs, ipairs
-local min = math.min,
-local type = type
-local t_concat = table.concat
-local t_insert = table.insert
-local select = select
-
-------------------------------------------------------------------
--- Arithmetic metamethods for strings
-------------------------------------------------------------------
-
-local mtstr = getmetatable('')
-function mtstr.__add(a,b) return a .. b end
-function mtstr.__mul(a,b) return type(a) == 'string' and a:rep(b) or b:rep(a) end
-function mtstr.__pow(a,b) return a:rep(b) end
-function mtstr.__div(a,b) return _.chop(a,b) end
-function mtstr.__unm(a) return a:reverse() end
-function mtstr.__mod(a,b)
-	local _chopped = _.chop(a,b)
-	return ((#a)%b==0) and nil or _chopped[#_chopped]
-end
-
-------------------------------------------------------------------
--- Private Helpers
-------------------------------------------------------------------
-
--- Creates a zero-matrix
-local function matrix(row,col)
-	local m = {}
-	for i = 1,row do m[i] = {}
-		for j = 1,col do m[i][j] = 0 end
-	end
-	return m
-end
-
--- Returns the minimum of a collection of numbers
-local function minOf(...)
-	local _arg = {...}
-	local m = _arg[1]
-	for k,v in ipairs(_arg) do m = min(m,v) end
-	return m
-end
-
--- Retuns the precedent/next character in Ascii table
-local function preOrSucc(char,dir)
-	return string.char(char:byte()+dir)
-end
-
--- Returns a list of functions keys packed into an object
-local function functions(obj)
-	local funcs = {}
-	for k,v in pairs(obj) do
-		if type(v) == 'function' then t_insert(funcs,k) end
-	end
-	return funcs
-end
-
--- Returns the index of value v in table t
-local function findValue(t,v)
-	for k,_v in pairs(t) do
-		if (_v==v) then return k end
-	end
-	return nil
-end
-
-------------------------------------------------------------------
 -- Library functions
 ------------------------------------------------------------------
 
@@ -108,7 +37,18 @@ end
 -- _.capitalizeFirst('uncapitalized') --> 'Uncapitalized' <br/>
 -- _.capitalizeFirst('Capitalized') -->  'Capitalized'
 -- </pre>
+-- @see _.capFirst
 function _.capitalizeFirst(str) end
+
+--- Capitalizes the first character of a given string. Alias for <tt>_.capitalizeFirst</tt>
+-- @param str a string
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.capFirst('uncapitalized') --> 'Uncapitalized' <br/>
+-- _.capFirst('Capitalized') -->  'Capitalized'
+-- </pre>
+-- @see _.capitalizeFirst
+function _.capFirst(str) end
 
 --- Capitalizes substring i to j in a given string
 -- @param str a string
@@ -118,9 +58,53 @@ function _.capitalizeFirst(str) end
 -- @usage <pre class='example'> 
 -- _.capitalize('capitalized') --> 'CAPITALIZED' <br/>
 -- _.capitalize('capitalized',5,9) --> 'capiTALIZed' <br/>
--- print(_.capitalize('capitalized',-8) --> 'capITALIZED'
+-- _.capitalize('capitalized',-8) --> 'capITALIZED'
 -- </pre>
+-- @see _.cap
 function _.capitalize(str,i,j) end
+
+--- Capitalizes substring i to j in a given string. Alias for <tt>_.capitalize</tt>
+-- @param str a string
+-- @param i a string. Defaults to 1 if not given. Can be negative (counting from right to left)
+-- @param j a string. Default to string length if not given. (counting from right to left)
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.cap('capitalized') --> 'CAPITALIZED' <br/>
+-- _.cap('capitalized',5,9) --> 'capiTALIZed' <br/>
+-- _.cap('capitalized',-8) --> 'capITALIZED'
+-- </pre>
+-- @see _.capitalize
+function _.cap(str,i,j) end
+
+--- Capitalizes each word in a given string.
+-- @param str a string
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.capitalizesEach('my tailor is rich') --> 'My Tailor Is Rich'
+-- </pre>
+-- @see _.capEach
+-- @see _.caps
+function _.capitalizesEach(str) end
+
+--- Capitalizes each word in a given string. Alias for <tt>_.capitalizesEach</tt>
+-- @param str a string
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.capEach('my tailor is rich') --> 'My Tailor Is Rich'
+-- </pre>
+-- @see _.capitalizesEach
+-- @see _.caps
+function _.capEach(str) end
+
+--- Capitalizes each word in a given string. Alias for <tt>_.capitalizesEach</tt>
+-- @param str a string
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.caps('my tailor is rich') --> 'My Tailor Is Rich'
+-- </pre>
+-- @see _.capitalizesEach
+-- @see _.capEach
+function _.caps(str) end
 
 --- Lowers the first character case in a given string
 -- @param str a string
@@ -142,17 +126,29 @@ function _.lowerFirst(str) end
 -- </pre>
 function _.lower(str,i,j) end
 
---- Tests if a given string contains any upper-case character
+--- Tests if a given string contains only upper-case characters
 -- @param str a string
 -- @return a string
 -- @usage <pre class='example'> 
 -- _.isLower('String') --> false <br/>
 -- _.isLower('strinG') --> false <br/>
--- print(_.isLower('string') --> true
+-- _.isLower('string') --> true
 -- </pre>
+-- @see _.isLowerCase
 function _.isLower(str) end
 
---- Tests if a given string contains any lower-case character
+--- Tests if a given string contains only upper-case characters. Alias for <tt>_.isLower</tt>
+-- @param str a string
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.isLowerCase('String') --> false <br/>
+-- _.isLowerCase('strinG') --> false <br/>
+-- _.isLowerCase('string') --> true
+-- </pre>
+-- @see _.isLower
+function _.isLowerCase(str) end
+
+--- Tests if a given string contains only lower-case characters
 -- @param str a string
 -- @return a string
 -- @usage <pre class='example'> 
@@ -160,7 +156,19 @@ function _.isLower(str) end
 -- _.isUpper('sTRinG') --> false <br/>
 -- _.isUpper('STRING') --> true
 -- </pre>
+-- @see _.isUpperCase
 function _.isUpper(str) end
+
+--- Tests if a given string contains only lower-case characters. Alias for <tt>_.isUpper</tt>
+-- @param str a string
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.isUpperCase('String') --> false <br/>
+-- _.isUpperCase('sTRinG') --> false <br/>
+-- _.isUpperCase('STRING') --> true
+-- </pre>
+-- @see _.isUpper
+function _.isUpperCase(str) end
 
 --- Tests if a given string starts with a lower-case character
 -- @param str a string
@@ -169,16 +177,38 @@ function _.isUpper(str) end
 -- _.startsLower('String') --> false <br/>
 -- _.startsLower('sTRinG') --> true
 -- </pre>
+-- @see _.startsLowerCase
 function _.startsLower(str) end
+
+--- Tests if a given string starts with a lower-case character. Alias for <tt>_.startsLower</tt>
+-- @param str a string
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.startsLowerCase('String') --> false <br/>
+-- _.startsLowerCase('sTRinG') --> true
+-- </pre>
+-- @see _.startsLower
+function _.startsLowerCase(str) end
 
 --- Tests if a given string starts with an upper-case character
 -- @param str a string
 -- @return a string
 -- @usage <pre class='example'> 
--- _.startsLower('String') --> false <br/>
--- _.startsLower('sTRinG') --> true
+-- _.startsUpper('String') --> true <br/>
+-- _.startsUpper('sTRinG') --> false
 -- </pre>
+-- @see _.startsUpperCase
 function _.startsUpper(str) end
+
+--- Tests if a given string starts with an upper-case character. Alias for <tt>_.startsUpper</tt>
+-- @param str a string
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.startsUpperCase('String') --> true <br/>
+-- _.startsUpperCase('sTRinG') --> false
+-- </pre>
+-- @see _.startsUpper
+function _.startsUpperCase(str) end
 
 --- Swaps the case of each characters in substring i-j inside a given string
 -- @param str a string
@@ -192,7 +222,109 @@ function _.startsUpper(str) end
 -- </pre>
 function _.swapCase(str,i,j) end
 
---- Returns the Levenshtein distance between stringA to stringB (See <a href="http://en.wikipedia.org/wiki/Levenshtein_distance" id="Levenshtein Distance on Wikipedia">Levenshtein Distance on Wikipedia</a>)
+--- Checks if a given string contains only alphabetic characters
+-- @param str a string
+-- @return a boolean
+-- @usage <pre class='example'> 
+-- _.isAlpha('text') --> true <br/>
+-- _.isAlpha('text1')  --> false <br/>
+-- _.isAlpha('te xt~')  --> false
+-- </pre>
+function _.isAlpha(str) end
+
+--- Checks if a given string contains only alphabetic characters
+-- @param str a string
+-- @return a boolean
+-- @usage <pre class='example'> 
+-- _.isNumeric('1e20')--> true <br/>
+-- _.isNumeric('-33.2a') --> false
+-- </pre>
+-- @see _.isNum
+function _.isNumeric(str) end
+
+--- Checks if a given string contains only alphabetic characters. Alias for <tt>_.isNumeric</tt>
+-- @param str a string
+-- @return a boolean
+-- @usage <pre class='example'> 
+-- _.isNum('1e20')--> true <br/>
+-- _.isNum('-33.2a') --> false
+-- </pre>
+-- @see _.isNumeric
+function _.isNum(str) end
+
+--- Checks if a given string contains only alphabetic and digits
+-- @param str a string
+-- @return a boolean
+-- @usage <pre class='example'> 
+-- _.isAlphaNumeric('-33.2a') --> false <br/>
+-- _.isAlphaNumeric('33.2') --> true
+-- </pre>
+-- @see _.isANum
+function _.isAlphaNumeric(str) end
+
+--- Checks if a given string contains only alphabetic and digits. Alias for <tt>_.isAlphaNumeric</tt>
+-- @param str a string
+-- @return a boolean
+-- @usage <pre class='example'> 
+-- _.isANum('-33.2a') --> false <br/>
+-- _.isANum('33.2') --> true
+-- </pre>
+-- @see _.isAlphaNumeric
+function _.isANum(str) end
+
+--- Checks if a given string can be converted to hex
+-- @param str a string
+-- @return a boolean
+-- @usage <pre class='example'> 
+-- _.isHexadecimal('-33.2') --> false <br/>
+-- _.isHexadecimal('FDA') --> true
+-- </pre>
+-- @see _.isHex
+function _.isHexadecimal(str) end
+
+--- Checks if a given string can be converted to hex. Alias for <tt>_.isHexadecimal</tt>
+-- @param str a string
+-- @return a boolean
+-- @usage <pre class='example'> 
+-- _.isHex('-33.2') --> false <br/>
+-- _.isHex('FDA') --> true
+-- </pre>
+-- @see _.isHexadecimal
+function _.isHex(str) end
+
+--- Checks if a given string matches a valid email address syntax. See <a href="http://en.wikipedia.org/wiki/Email_address" id="Email Addresses on Wikipedia">Email Addresses on Wikipedia</a>
+-- @param str a string
+-- @return a boolean
+-- @usage <pre class='example'> 
+-- _.isEmail('@') --> false <br/>
+-- _.isEmail('.com') --> false <br/>
+-- _.isEmail('a@com') --> false <br/>
+-- _.isEmail('~user..nm.@web.co.uk') --> false <br/>
+-- _.isEmail('user.nm@web.co.uk') --> true 
+-- </pre>
+function _.isEmail(str) end
+
+--- Indexes a string like a table, returning a character. 
+-- @param str a string
+-- @return i an index
+-- @usage <pre class='example'> 
+-- _.index('Hello',2) --> 'e'  <br/>
+-- _.index('Hello',10) --> nil
+-- </pre>
+-- @see _.charAt
+function _.index(str) end
+
+--- Indexes a string like a table, returning a character. Alias for <tt>_.index</tt>
+-- @param str a string
+-- @return i an index
+-- @usage <pre class='example'> 
+-- _.charAt('Hello',2) --> 'e'  <br/>
+-- _.charAt('Hello',10) --> nil
+-- </pre>
+-- @see _.index
+function _.charAt(str) end
+
+--- Returns the Levenshtein distance between stringA to stringB. See <a href="http://en.wikipedia.org/wiki/Levenshtein_distance" id="Levenshtein Distance on Wikipedia">Levenshtein Distance on Wikipedia</a>
 -- @param strA a string
 -- @param strB another string
 -- @return a number
@@ -238,8 +370,18 @@ function _.includes(str,sub) end
 -- @usage <pre class='example'> 
 -- _.chars('array') --> {'a','r','r','a','y'}
 -- </pre>
+-- @see _.explode
 -- @see _.chop
 function _.chars(str) end
+
+--- Converts a given string to an array of chars. Alias for <tt>_.chars</tt>
+-- @param str a string
+-- @return an array
+-- @usage <pre class='example'> 
+-- _.explode('array') --> {'a','r','r','a','y'}
+-- </pre>
+-- @see _.chars
+function _.explode(str) end
 
 --- Returns the number of substring occurences in a given string
 -- @param str a string
@@ -330,7 +472,20 @@ function _.endsWith(str,ends) end
 -- _.succ('wxy') --> 'xyz' <br/>
 -- _.succ('D',3) --> 'G'
 -- </pre>
+-- @see _.next
 function _.succ(str,step) end
+
+--- Returns the successor of a given character or set of characters. Alias for <tt>_.succ</tt>
+-- @param str a string
+-- @param step If given, returns the step-th successor of the given string. Defaults to 1 if not given.
+-- @return a string
+-- @usage <pre class='example'> 
+-- _.next('a') --> 'b' <br/>
+-- _.next('wxy') --> 'xyz' <br/>
+-- _.next('D',3) --> 'G'
+-- </pre>
+-- @see _.succ
+function _.next(str,step) end
 
 --- Returns the predecessor of a given character or set of characters
 -- @param str a string