+++ version_history.md
@@ -1,4 +1,15 @@
 #Version history#
 
+##1.0 (08/02/12)
+* Added <tt>_.isAlpha</tt>, <tt>_.isNumeric</tt>, <tt>_.isAlphaNumeric</tt> and their aliases
+* Added <tt>_.isHexadecimal</tt> and <tt>_.isHex</tt> alias
+* Added <tt>_.startsLowerCase</tt>, <tt>_.startsUpperCase</tt> aliases
+* Added <tt>_.isLowerCase</tt>, <tt>_.isUpperCase</tt> aliases
+* Added <tt>_.explode</tt>, <tt>_.cap</tt>, <tt>_.capFirst</tt>, <tt>_.next</tt> aliases
+* Added <tt>_.capitalizesEach</tt> and <tt>_.capEach</tt>, <tt>_.caps</tt> aliases
+* Added string indexing feature
+* Updated documentation
+* Updated Allen_Lib_Test.lua and Allen_StringOp_Tests.lua
+
 ##0.1 (07/24/12)
 * Initial Release